requirements.txt
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
Django==5.1.7
django-browser-reload==1.18.0
django-crispy-forms==2.3
django-tailwind==3.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.7
numpy==2.3.0
packaging==24.2
pandas==2.3.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
whitenoise==6.9.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
